requirements.txt
fastapi
uvicorn
qdrant-client
openai
PyGithub
httpx
python-dotenv
tiktoken
pydantic
